评论页显示头像

+++ LXSBlog-framework/src/main/java/com/lxs/utils/RedisCache.java
@@ -243,7 +243,7 @@ public class RedisCache
      * @param hKey h键
      * @param v    v
      */
-    public void incrementCacheMapValue(String key,String hKey,long v){
+    public void incrementCacheMapValue(String key,String hKey,int v){
         redisTemplate.opsForHash().increment(key,hKey,v);
     }
 }
